Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/bad-strategies.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL bad-strategies.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
